# docs/swagger.yaml
basePath: /
definitions:
  models.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  models.User:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      profile:
        type: string
    type: object
info:
  contact: {}
paths:
  /login:
    post:
      consumes:
      - multipart/form-data
      description: Melakukan login user menggunakan email dan password
      parameters:
      - description: Email user
        in: formData
        name: email
        required: true
        type: string
      - description: Password user
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Login berhasil
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Email atau password salah
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: User tidak ditemukan
          schema:
            $ref: '#/definitions/models.Response'
      summary: Login user
      tags:
      - Auth
  /register:
    post:
      consumes:
      - multipart/form-data
      description: Mendaftarkan user baru menggunakan name, email, dan password
      parameters:
      - description: Nama lengkap
        in: formData
        name: name
        required: true
        type: string
      - description: Email user
        in: formData
        name: email
        required: true
        type: string
      - description: Password minimal 8 karakter
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Registrasi berhasil
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Input tidak valid
          schema:
            $ref: '#/definitions/models.Response'
      summary: Register user baru
      tags:
      - Auth
  /update/profile/{id}:
    patch:
      consumes:
      - multipart/form-data
      description: Update data user (name, email, password) dan upload foto profil
        (single image upload)
      parameters:
      - description: Foto profil user (image)
        in: formData
        name: pic
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Upload dan update profile user
      tags:
      - Users
  /users:
    get:
      description: Mengambil semua data user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Membuat user baru
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Create new user
      tags:
      - Users
  /users/{id}:
    delete:
      description: Menghapus user berdasarkan ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete user by ID
      tags:
      - Users
    get:
      description: Mengambil data user berdasarkan ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Mengupdate data user berdasarkan ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update user by ID
      tags:
      - Users
swagger: "2.0"